+++ myCookingManager/frontend/src/components/RecipeDetailsPage/RecipeDetailsCategorySelect.jsx
@@ -4,7 +4,7 @@ import { useAuth0 } from "@auth0/auth0-react";
 
 import { RecipeDetailsContext } from "../contexts/RecipeDetailsContext";
 
-import CategoryCreation from "../CategoryCreation";
+import CategoryCreation from "../multipage_components/CategoryCreation";
 
 const RecipeDetailsCategorySelect = () => {
 	//Import user object from auth0
