+++ myCookingManager/frontend/src/components/multipage_components/RatingSystem.jsx
@@ -6,8 +6,8 @@ import { styled } from "styled-components";
 import { Rating } from "@smastrom/react-rating";
 import "@smastrom/react-rating/style.css";
 
-import { RecipeDetailsContext } from "./contexts/RecipeDetailsContext";
-import { allRatingsCalculator } from "../utility_functions/allRatingsCalculator";
+import { RecipeDetailsContext } from "../contexts/RecipeDetailsContext";
+import { allRatingsCalculator } from "../../utility_functions/allRatingsCalculator";
 
 const RatingSystem = () => {
 	// Import user object from auth0
